chore: seperate updater tests into 2 files
one file for general updater tests, one file for
hook tests
